ref(ConnectWalletModal): cleanup
