+++ components/common/ConnectWallet/ConnectWalletModal/Evm.vue
@@ -1,17 +1,19 @@
 <template>
-  <NeoButton
-    variant="k-accent"
-    no-shadow
-    size="large"
-    class="w-full"
-    @click="openModal">
-    Connect with Web3Modal
-  </NeoButton>
+  <div>
+    <NeoButton
+      variant="k-accent"
+      no-shadow
+      size="large"
+      class="w-full"
+      @click="openModal">
+      Connect with Web3Modal
+    </NeoButton>
+  </div>
 </template>
 <script lang="ts" setup>
 import { NeoButton } from '@kodadot1/brick'
 
-const emits = defineEmits(['setAccount'])
+const emits = defineEmits(['select'])
 
 const { openModal, address } = useWeb3Modal()
 const walletStore = useWalletStore()
@@ -22,7 +24,7 @@ watch(address, (address) => {
       address: address as string,
       vm: 'EVM',
     })
-    emits('setAccount', { address })
+    emits('select', { address })
   }
 })
 </script>
